# docs/api/swagger.yaml
components:
  schemas:
...
    ApiError:
      properties:
        code:
          example: BAD_REQUEST
          type: string
        details:
          description: |-
            Details contains additional error context (field names, validation rules, etc.)
            Type varies by error code: map[string]string for validation, string for general errors
          example: '{"field":"email","rule":"required"}'
        message:
          example: Invalid request parameters
          type: string
      type: object
...
    DependencyStatus:
      properties:
        id:
          type: string
        isCompleted:
          type: boolean
        status:
          type: string
      type: object
...
    ExecutionResult:
      properties:
        duration:
          description: Execution duration in milliseconds
          type: integer
        output:
          description: Execution output/logs
          type: string
        success:
          description: Whether execution succeeded
          type: boolean
        testResults:
          description: Test results (if applicable)
          items:
            $ref: '#/components/schemas/TestResult'
          type: array
          uniqueItems: false
      type: object
...
    RetryTask:
      properties:
        force:
          description: Force retry even if dependencies not met
          example: false
          type: boolean
      type: object
...
    SkipTask:
      properties:
        reason:
          example: not_applicable
          type: string
      required:
      - reason
      type: object
    Success:
      properties:
        message:
          example: Operation completed
          type: string
        status:
          example: success
          type: string
      type: object
    Task:
      properties:
        canExecute:
          example: true
          type: boolean
        dependencies:
          example:
          - '["task-id-1"]'
          items:
            type: string
          type: array
          uniqueItems: false
        description:
          example: Implement login logic
          type: string
        failureReason:
          example: Test failed
          type: string
        id:
          example: 550e8400-e29b-41d4-a716-446655440000
          type: string
        order:
          example: 1
          type: integer
        retryCount:
          example: 0
          type: integer
        status:
          example: completed
          type: string
        statusDisplay:
          example: Completed
          type: string
        suggestion:
          example: Fix the test
          type: string
      type: object
    TaskList:
      properties:
        allCompleted:
          type: boolean
        completed:
          type: integer
        failed:
          type: integer
        hasFailedTask:
          type: boolean
        inProgress:
          type: integer
        maxRetryLimit:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        pending:
          type: integer
        sessionId:
          type: string
        skipped:
          type: integer
        tasks:
          items:
            $ref: '#/components/schemas/Task'
          type: array
          uniqueItems: false
        totalItems:
          description: Pagination info (based on filtered results)
          type: integer
        totalPages:
          type: integer
        totalTasks:
          description: Total tasks in session (unfiltered)
          type: integer
      type: object
    TaskStatus:
      properties:
        dependencies:
          items:
            $ref: '#/components/schemas/DependencyStatus'
          type: array
          uniqueItems: false
        description:
          example: Implement login logic
          type: string
        executionResult:
          $ref: '#/components/schemas/ExecutionResult'
        failureReason:
          example: Test failed
          type: string
        order:
          example: 1
          type: integer
        retryCount:
          example: 0
          type: integer
        sessionId:
          example: 550e8400-e29b-41d4-a716-446655440000
          type: string
        status:
          example: completed
          type: string
        statusDisplayName:
          example: Completed
          type: string
        suggestion:
          example: Fix the test
          type: string
        taskId:
          example: 660e8400-e29b-41d4-a716-446655440001
          type: string
      type: object
...
    TestResult:
      properties:
        message:
          description: Error message (if failed)
          type: string
        name:
          description: Test name
          type: string
        status:
          description: passed, failed, skipped
          type: string
      type: object
...
info:
  description: Automated development agent system - from GitHub Issue to Pull Request
  title: Litchi API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/sessions/{sessionId}/tasks:
    get:
      description: Retrieve all tasks for a work session with status statistics and
        execution progress
      parameters:
      - description: Session ID (UUID format)
        in: path
        name: sessionId
        required: true
        schema:
          type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        schema:
          minimum: 1
          type: integer
      - description: 'Page size (default: 20, max: 100)'
        in: query
        name: pageSize
        schema:
          maximum: 100
          minimum: 1
          type: integer
      - description: Filter by status (pending, in_progress, completed, failed, skipped)
        in: query
        name: status
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
          description: Task list retrieved successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Invalid session ID format
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Session not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: Get task list
      tags:
      - tasks
  /api/v1/sessions/{sessionId}/tasks/{taskId}:
    get:
      description: Retrieve detailed status information for a specific task including
        dependencies and execution result
      parameters:
      - description: Session ID (UUID format)
        in: path
        name: sessionId
        required: true
        schema:
          type: string
      - description: Task ID (UUID format)
        in: path
        name: taskId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatus'
          description: Task status retrieved successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Invalid session ID or task ID format
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Session or task not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: Get task status
      tags:
      - tasks
  /api/v1/sessions/{sessionId}/tasks/{taskId}/retry:
    post:
      description: Reset a failed task to in-progress status for re-execution. Task
        must be in failed status and not exceed retry limit.
      parameters:
      - description: Session ID (UUID format)
        in: path
        name: sessionId
        required: true
        schema:
          type: string
      - description: Task ID (UUID format)
        in: path
        name: taskId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/RetryTask'
                description: Retry options
                summary: body
        description: Retry options
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
          description: Task retry initiated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Invalid request body, session ID, or task ID
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Session or task not found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Task cannot be retried (not failed or exceeded retry limit)
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: Retry task
      tags:
      - tasks
  /api/v1/sessions/{sessionId}/tasks/{taskId}/skip:
    post:
      description: Mark a task as skipped, removing it from execution queue. Skipped
        tasks are considered completed for dependency resolution.
      parameters:
      - description: Session ID (UUID format)
        in: path
        name: sessionId
        required: true
        schema:
          type: string
      - description: Task ID (UUID format)
        in: path
        name: taskId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/SkipTask'
                description: Skip reason
                summary: body
        description: Skip reason
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
          description: Task skipped successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Invalid request body, session ID, or task ID
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Session or task not found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Task cannot be skipped (invalid status)
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: Skip task
      tags:
      - tasks
...